+++ CHANGELOG.md
@@ -9,6 +9,16 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [0.3.0] - 2020-12-14
+
+### Changed
+
+- Changed `CompatiblePairs` to a function that accepts a list of products
+
+### Removed
+
+- Remove `DriversByProduct` and `ProductsByDriver`
+
 ## [0.2.0] - 2020-12-14
 
 ### Added
@@ -23,6 +33,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Unreleased]: https://github.com/dogmatiq/sqltest
 [0.1.0]: https://github.com/dogmatiq/sqltest/releases/tag/v0.1.0
 [0.2.0]: https://github.com/dogmatiq/sqltest/releases/tag/v0.2.0
+[0.3.0]: https://github.com/dogmatiq/sqltest/releases/tag/v0.3.0
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD